181494950 rewrite api controller (#384)
* ValidatorQuery and api Validator Controller

* validators index and show

* passing tests for validators api

* passing tests

* fix quotes

* add permitted attributes

* fix quotes

* passing tests

* eager load -> joins

* Update app/controllers/api/v1/validator_block_histories_controller.rb


* Update app/queries/validator_query.rb


* Update test/queries/validator_query_test.rb


* Update test/queries/validator_query_test.rb


* Update test/queries/validator_query_test.rb


* Update test/queries/validator_query_test.rb


* fix quotes

* global networks in test
