+++ app/controllers/api/v1/api_controller.rb
@@ -25,96 +25,6 @@ module Api
         render json: { 'status' => 'Method Gone' }, status: 410
       end
 
-      # Show the list of validators with scores
-      def validators_list
-        @sort_order = case params[:order]
-                      when 'score'
-                        'validator_score_v1s.total_score desc,  validator_score_v1s.active_stake desc'
-                      when 'name'
-                        'validators.name asc'
-                      when 'stake'
-                        'validator_score_v1s.active_stake desc, validator_score_v1s.total_score desc'
-                      else
-                        'RAND()'
-                      end
-
-        limit = params[:limit] || 9999
-        page = params[:page]
-
-        @validators = Validator.select(validator_fields, validator_score_v1_fields)
-                               .where(network: params[:network])
-                               .includes(:vote_accounts, :most_recent_epoch_credits_by_account, validator_score_v1: [:ip_for_api])
-                               .joins(:validator_score_v1)
-                               .order(@sort_order)
-                               .page(page)
-                               .per(limit)
-
-        unless params[:q].blank?
-          @validators = ValidatorSearchQuery.new(@validators).search(params[:q])
-        end
-
-        @skipped_slots_report = Report.where(
-          network: params[:network],
-          name: 'build_skipped_slot_percent'
-        ).last
-
-        json_result = @validators.map { |val| create_json_result(val) }
-
-        render json: json_result
-      rescue ActionController::ParameterMissing
-        render json: { 'status' => 'Parameter Missing' }, status: 400
-      rescue StandardError => e
-        Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
-      end
-
-      def validators_show
-        @validator = Validator.select(validator_fields, validator_score_v1_fields)
-                              .eager_load(:validator_score_v1)
-                              .find_by(network: params[:network], account: params['account'])
-
-        raise ValidatorNotFound if @validator.nil?
-
-        @skipped_slots_report = Report.where(
-          network: params[:network],
-          name: 'build_skipped_slot_percent'
-        ).last
-
-        current_epoch = EpochHistory.last
-
-        render json: create_json_result(@validator, params[:with_history] || false)
-      rescue ValidatorNotFound
-        render json: { 'status' => 'Validator Not Found' }, status: 404
-      rescue ActionController::ParameterMissing
-        render json: { 'status' => 'Parameter Missing' }, status: 400
-      rescue StandardError => e
-        Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
-      end
-
-      def validator_block_history
-        @validator = Validator.where(
-          network: params[:network], account: params['account']
-        ).order('network, account').first
-
-        raise ValidatorNotFound if @validator.nil?
-
-        @limit = params[:limit] || 9999
-
-        @block_history = @validator.validator_block_histories
-                                   .order('id desc')
-                                   .limit(@limit)
-
-        render 'api/v1/validators/block_history', formats: :json
-      rescue ValidatorNotFound
-        render json: { 'status' => 'Validator Not Found' }, status: 404
-      rescue ActionController::ParameterMissing
-        render json: { 'status' => 'Parameter Missing' }, status: 400
-      rescue StandardError => e
-        Appsignal.send_error(e)
-        render json: { 'status' => e.message }, status: 500
-      end
-
       # Filter params for the Collector
       def collector_params
         params.require(:collector).permit(
@@ -122,80 +32,6 @@ module Api
           :payload_version,
           :payload
         )
-      end
-
-      private
-
-      def create_json_result(validator, with_history = false)
-        hash = {}
-
-        hash.merge!(validator.to_builder.attributes!)
-
-        score = validator.score
-        ip = score.ip_for_api if score
-        vote_account = validator.vote_accounts.last
-        validator_history = validator.most_recent_epoch_credits_by_account
-
-        
-        hash.merge!(score.to_builder(with_history: with_history).attributes!)
-        hash.merge!(ip.to_builder.attributes!) unless ip.blank?
-        hash.merge!(vote_account.to_builder.attributes!) unless vote_account.blank?
-        hash.merge!(validator_history.to_builder.attributes!) unless validator_history.blank?
-
-        # Data from the skipped_slots_report
-        unless @skipped_slots_report.nil?
-          this_report = @skipped_slots_report.payload.select do |ssr|
-            ssr['account'] == validator.account
-          end
-
-          unless this_report.first.nil?
-            hash.merge!({
-              'skipped_slots' => this_report.first['skipped_slots'],
-              'skipped_slot_percent' => this_report.first['skipped_slot_percent'],
-              'ping_time' => nil # this_report.first['ping_time']
-            })
-          end
-        end
-
-        hash.merge!({url: validator.api_url})
-        hash
-      end
-
-      def validator_fields
-        [
-          'account',
-          'created_at',
-          'details',
-          'id',
-          'keybase_id',
-          'name',
-          'network',
-          'updated_at',
-          'www_url',
-          'avatar_url',
-          'admin_warning'
-        ].map { |e| "validators.#{e}" }
-      end
-
-      def validator_score_v1_fields
-        [
-          'active_stake',
-          'commission',
-          'delinquent',
-          'data_center_concentration_score',
-          'data_center_key',
-          'data_center_host',
-          'published_information_score',
-          'root_distance_score',
-          'security_report_score',
-          'skipped_slot_score',
-          'software_version',
-          'software_version_score',
-          'stake_concentration_score',
-          'total_score',
-          'validator_id',
-          'vote_distance_score'
-        ].map { |e| "validator_score_v1s.#{e}" }
       end
     end
   end
